[Comments] Refactor quill, bounty functions into common modules
