+++ components/Comment/CommentEditor.tsx
@@ -4,8 +4,7 @@ import { css, StyleSheet } from "aphrodite";
 import { useEffect, useRef, useState } from "react";
 import Button from "../Form/Button";
 import CreateBountyBtn from "../Bounty/CreateBountyBtn";
-import buildQuillModules from "./lib/buildQuillModules";
-import QuillFormats from "./lib/quillFormats";
+import { QuillFormats, buildQuillModules } from "./lib/quill";
 import isQuillEmpty from "../TextEditor/util/isQuillEmpty";
 import { AuthorProfile } from "~/config/types/root_types";
 import CommentAuthors from "./CommentAuthors";
