--- components/Comment/lib/isQuillEmpty.ts
@@ -1,3 +1,0 @@
-export default function isQuillEmpty(content) {
-  return !content || JSON.stringify(content) === '{"ops":[{"insert":"\\n"}]}';
-}
