+++ components/Comment/lib/quill.ts
@@ -1,0 +1,50 @@
+export const buildQuillModules = ({ editorId, handleSubmit, handleImageUpload }) => {
+  const modules = {
+    // magicUrl: true,
+    keyboard: {
+      bindings: {
+        commandEnter: {
+          key: 13,
+          shortKey: true,
+          metaKey: true,
+          handler: handleSubmit,
+        },
+      },
+    },
+    toolbar: {
+      // magicUrl: true,
+      container: `#${editorId}`,
+      handlers: {
+        image: handleImageUpload,
+      },
+    },
+  };
+
+  return modules;
+}
+
+export default function isQuillEmpty(content) {
+  return !content || JSON.stringify(content) === '{"ops":[{"insert":"\\n"}]}';
+}
+
+export const QuillFormats = [
+  "image",
+  "header",
+  "font",
+  "size",
+  "bold",
+  "italic",
+  "underline",
+  "strike",
+  "blockquote",
+  "list",
+  "bullet",
+  "indent",
+  "link",
+  "video",
+  "clean",
+  "background",
+  "code-block",
+  "direction",
+  // "peer-review-rating",
+];
