--- components/Comment/lib/quillFormats.ts
@@ -1,23 +1,0 @@
-const QuillFormats = [
-  "image",
-  "header",
-  "font",
-  "size",
-  "bold",
-  "italic",
-  "underline",
-  "strike",
-  "blockquote",
-  "list",
-  "bullet",
-  "indent",
-  "link",
-  "video",
-  "clean",
-  "background",
-  "code-block",
-  "direction",
-  // "peer-review-rating",
-];
-
-export default QuillFormats;
